Crecard #03

+++ src/br/com/mind5/payment/creditCard/info/CrecardInfo.java
@@ -13,7 +13,7 @@ public final class CrecardInfo extends InfoRecord implements Cloneable {
 	public long codOwner;
 	public long codCreditCard;
 	public long codPayCustomer;
-	public long codPayPartner;
+	public int codPayPartner;
 	public long codUser;
 	public String creditCardId;
 	public String creditCardBrand;
